add further edge case tests for pagination query parameters

+++ api/src/main/scala/uk/ac/wellcome/platform/api/models/DisplaySearch.scala
@@ -14,7 +14,7 @@ case object DisplaySearch {
     DisplaySearch(
       results = searchResponse.hits.map { DisplayWork(_) },
       pageSize = pageSize,
-      totalPages = Math.ceil(searchResponse.totalHits.toInt / pageSize).toInt,
+      totalPages = Math.ceil(searchResponse.totalHits.toDouble / pageSize.toDouble).toInt,
       totalResults = searchResponse.totalHits.toInt
     )
   }
